refactor: tool button
